+++ src/authz-module/roles-permissions/course/constants.ts
@@ -1,4 +1,6 @@
-import { PermissionMetadata, ResourceMetadata, RoleMetadata } from '@src/types';
+import {
+  PermissionMetadata, ResourceMetadata, Role, RoleMetadata,
+} from '@src/types';
 import {
   LibraryBooks, Article, Group, LocalOffer,
   BookOpen,
@@ -298,10 +300,11 @@ export const coursePermissions: PermissionMetadata[] = [
 ];
 
 // roles hardcoded, todo: need to add the constants from above in order to merge the different permissions array.
-export const rolesObject = [
+export const rolesObject: Role[] = [
   {
     role: 'course_admin',
     contextType: 'course',
+    scope: '',
     permissions: [
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE,
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE_UPDATES,
@@ -342,6 +345,7 @@ export const rolesObject = [
   {
     role: 'course_staff',
     contextType: 'course',
+    scope: '',
     permissions: [
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE,
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE_UPDATES,
@@ -379,6 +383,7 @@ export const rolesObject = [
   {
     role: 'course_editor',
     contextType: 'course',
+    scope: '',
     permissions: [
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE,
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE_UPDATES,
@@ -409,6 +414,7 @@ export const rolesObject = [
   {
     role: 'course_auditor',
     contextType: 'course',
+    scope: '',
     permissions: [
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE,
       CONTENT_COURSE_PERMISSIONS.VIEW_COURSE_UPDATES,
